Finances component extracted to stand-alone file
